+++ components/SEOBundle/bundle/GraphQL/Schema/Domain/Content/Mapper/FieldDefinition/NovaeZSEOFieldDefinitionMapper.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Novactive\Bundle\eZSEOBundle\GraphQL\Schema\Domain\Content\Mapper\FieldDefinition;
+
+use eZ\Publish\API\Repository\Values\ContentType\FieldDefinition;
+use EzSystems\EzPlatformGraphQL\Schema\Domain\Content\Mapper\FieldDefinition\DecoratingFieldDefinitionMapper;
+use EzSystems\EzPlatformGraphQL\Schema\Domain\Content\Mapper\FieldDefinition\FieldDefinitionMapper;
+
+class NovaeZSEOFieldDefinitionMapper extends DecoratingFieldDefinitionMapper implements FieldDefinitionMapper
+{
+    protected function getFieldTypeIdentifier(): string
+    {
+        return 'novaseometas';
+    }
+
+    public function mapToFieldValueType(FieldDefinition $fieldDefinition): ?string
+    {
+        if (!$this->canMap($fieldDefinition)) {
+            return parent::mapToFieldValueType($fieldDefinition);
+        }
+
+        return "NovaSeoMetasFieldValue";
+    }
+
+    public function mapToFieldValueResolver(FieldDefinition $fieldDefinition): ?string
+    {
+        if (!$this->canMap($fieldDefinition)) {
+            return parent::mapToFieldValueResolver($fieldDefinition);
+        }
+
+        return '@=resolver("NovaSeoMetasFieldValue", [value, "' . $fieldDefinition->identifier . '"])';
+    }
+
+}